app/requirements.txt
flask
pyjwt
requests
cryptography
prometheus_client
mysql-connector-python
pytest
pytest-cov